mkdir in tests

+++ tests/test_examples.py
@@ -14,8 +14,11 @@ def test_example_1():
     reference_directory = REFERENCE_DIRECTORY / 'test_example_1'
     output_directory = OUTPUT_DIRECTORY / 'example_1'
 
-    for filename in output_directory.iterdir():
-        os.remove(filename)
+    if not output_directory.exists():
+        output_directory.mkdir(exist_ok=True, parents=True)
+    else:
+        for filename in output_directory.iterdir():
+            os.remove(filename)
 
     exec(open(EXAMPLES_DIRECTORY / 'example_1.py').read())
 
@@ -28,8 +31,11 @@ def test_example_2():
     reference_directory = REFERENCE_DIRECTORY / 'test_example_2'
     output_directory = OUTPUT_DIRECTORY / 'example_2'
 
-    for filename in output_directory.iterdir():
-        os.remove(filename)
+    if not output_directory.exists():
+        output_directory.mkdir(exist_ok=True, parents=True)
+    else:
+        for filename in output_directory.iterdir():
+            os.remove(filename)
 
     exec(open(EXAMPLES_DIRECTORY / 'example_2.py').read())
 
@@ -42,8 +48,11 @@ def test_example_3():
     reference_directory = REFERENCE_DIRECTORY / 'test_example_3'
     output_directory = OUTPUT_DIRECTORY / 'example_3'
 
-    for filename in output_directory.iterdir():
-        os.remove(filename)
+    if not output_directory.exists():
+        output_directory.mkdir(exist_ok=True, parents=True)
+    else:
+        for filename in output_directory.iterdir():
+            os.remove(filename)
 
     exec(open(EXAMPLES_DIRECTORY / 'example_3.py').read())
 
@@ -56,8 +65,11 @@ def test_example_4():
     reference_directory = REFERENCE_DIRECTORY / 'test_example_4'
     output_directory = OUTPUT_DIRECTORY / 'example_4'
 
-    for filename in output_directory.iterdir():
-        os.remove(filename)
+    if not output_directory.exists():
+        output_directory.mkdir(exist_ok=True, parents=True)
+    else:
+        for filename in output_directory.iterdir():
+            os.remove(filename)
 
     exec(open(EXAMPLES_DIRECTORY / 'example_4.py').read())
